https://issues.apache.org/jira/browse/SPARK-1892
Adding the breeze update in the project file. This update fixes the problem with breeze's own OWL-QN implementation.

Adding OWL-QN optimizer for L1 regularizations. It can also handle L2 regularizations together. It extends LBFGS. It uses the OWL-QN implementation from breeze (which didn't work correctly before, but it was also fixed prior to this). It requires the latest version of breeze to work correctly.