https://issues.apache.org/jira/browse/SPARK-1892
Updating the pom.xml file to reflect the breeze version change.

Adding OWL-QN optimizer for L1 regularizations. It can also handle L2 regularizations together. It extends LBFGS. It uses the OWL-QN implementation from breeze (which didn't work correctly before, but it was also fixed prior to this). It requires the latest version of breeze to work correctly.